Feat: 맵 데이터 create, read all

+++ src/main/java/com/yogo/metacraft/test/document/InstanceData.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class InstanceData {
-    private int modelNum;
+    private Double modelNum;
     private String mapName;
     private String serialNumber;
     private Position position;
